move this enum out on it's own
