+++ app/src/lib/dispatcher/error-handlers.ts
@@ -11,7 +11,7 @@ import { Repository } from '../../models/repository'
 import { PopupType } from '../../lib/app-state'
 import { ShellError } from '../shells'
 import { UpstreamAlreadyExistsError } from '../stores/upstream-already-exists-error'
-import { FetchType } from '../stores/index'
+import { FetchType } from '../../models/fetch'
 
 /** An error which also has a code property. */
 interface IErrorWithCode extends Error {
