+++ app/src/models/fetch.ts
@@ -1,0 +1,8 @@
+/**
+ * Enum used by fetch to determine if
+ * a fetch was initiated by the backgroundFetcher
+ */
+export enum FetchType {
+  BackgroundTask,
+  UserInitiatedTask,
+}
